refactor(fsd): RequireAuth を app/ へ移し shared/theme の境界を lint で固定する (#507)
WS-05/04 の安全な部分。RequireAuth は authStore（entities/auth）に依存する認可ゲート UI で、shared/auth に置かれ shared→entities の上向き依存になっていた。app/（app→entities は許可）へ移動し shared/auth ディレクトリを解消。consumer は router.tsx の1件のみ。あわせて純トークン基盤である shared/theme→entities を import/no-restricted-paths に追加。

注: shared→entities は当初想定（5箇所）より広く、shared/lib の公開レコード系ドメインヘルパ6ファイル（bootstrap/display-label/field-format/seed-cache）と shared/api/client.ts の auth store が load-bearing で、これらの zone 化は当該ヘルパの entities/features への再配置という別個のアーキテクチャ refactor を要する。entity-tag→tag(sibling) の TagId 参照も consumer 跨ぎの brand 戦略判断を伴うため、いずれも #507 のフォローアップとして分離。
